Add files via upload
For this project, i proposed the idea of creating a Furniture E-commerce app which helps to harnesses the power of technology to benefit individuals. This app employs AI, machine learning (ML), and augmented reality (AR) to offer an engaging experience where users can scan furniture items and then view or imagine them within their own living spaces. This innovative approach replaces the conventional method of furniture shopping, making it more convenient. Additionally, it serves as a platform to enhance the visibility and accessibility of small-scale furniture retailers.